Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,5 +3,5 @@
 int main(){
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    printf("Average: ",a+b+c/3);
+    printf("Average: ",(float)(a+b+c)/3);
 }
